# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Progi CAP API",
        "description": "API documentation for Progi CAP application modules",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/submissions/active-periods": {
            "get": {
                "tags": [
                    "Submission Management"
                ],
                "summary": "Périodes de soumission actives",
                "description": "Récupère la liste des périodes de soumission actuellement actives",
                "operationId": "getActiveSubmissionPeriods",
                "responses": {
                    "200": {
                        "description": "Périodes récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SubmissionPeriod"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/submissions/active-reclamation-periods": {
            "get": {
                "tags": [
                    "Submission Management"
                ],
                "summary": "Périodes de réclamation actives",
                "description": "Récupère la liste des périodes de réclamation actuellement actives",
                "operationId": "getActiveReclamationPeriods",
                "responses": {
                    "200": {
                        "description": "Périodes récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReclamationPeriod"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/submissions/check-status": {
            "post": {
                "tags": [
                    "Submission Management"
                ],
                "summary": "Vérifier le statut de soumission",
                "description": "Vérifie si la soumission est ouverte pour une année académique donnée",
                "operationId": "checkSubmissionStatus",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "academic_year_id"
                                ],
                                "properties": {
                                    "academic_year_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut vérifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "is_open": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "academic_year": {
                                                    "$ref": "#/components/schemas/AcademicYear"
                                                },
                                                "submission_period": {
                                                    "$ref": "#/components/schemas/SubmissionPeriod"
                                                },
                                                "current_time": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/submissions/check-reclamation-status": {
            "post": {
                "tags": [
                    "Submission Management"
                ],
                "summary": "Vérifier le statut de réclamation",
                "description": "Vérifie si la réclamation est ouverte pour une année académique donnée",
                "operationId": "checkReclamationStatus",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "academic_year_id"
                                ],
                                "properties": {
                                    "academic_year_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut vérifié avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "is_open": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "academic_year": {
                                                    "$ref": "#/components/schemas/AcademicYear"
                                                },
                                                "reclamation_period": {
                                                    "$ref": "#/components/schemas/ReclamationPeriod"
                                                },
                                                "current_time": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/academic-years": {
            "get": {
                "tags": [
                    "Submission Management"
                ],
                "summary": "Liste des années académiques",
                "description": "Récupère la liste de toutes les années académiques",
                "operationId": "getAcademicYears",
                "responses": {
                    "200": {
                        "description": "Années académiques récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AcademicYear"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/academic-years/{academicYear}": {
            "get": {
                "tags": [
                    "Submission Management"
                ],
                "summary": "Détails d'une année académique",
                "description": "Récupère les détails d'une année académique spécifique avec ses périodes",
                "operationId": "getAcademicYear",
                "parameters": [
                    {
                        "name": "academicYear",
                        "in": "path",
                        "description": "ID de l'année académique",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Année académique récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AcademicYear"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Année académique non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AcademicYear": {
                "title": "Academic Year",
                "description": "Modèle représentant une année académique",
                "properties": {
                    "id": {
                        "description": "ID unique",
                        "type": "integer"
                    },
                    "submission_start": {
                        "description": "Date de début des soumissions",
                        "type": "string",
                        "format": "date-time"
                    },
                    "submission_end": {
                        "description": "Date de fin des soumissions",
                        "type": "string",
                        "format": "date-time"
                    },
                    "reclamation_start": {
                        "description": "Date de début des réclamations",
                        "type": "string",
                        "format": "date-time"
                    },
                    "reclamation_end": {
                        "description": "Date de fin des réclamations",
                        "type": "string",
                        "format": "date-time"
                    },
                    "academic_year": {
                        "description": "Année académique (ex: 2023-2024)",
                        "type": "string"
                    },
                    "year_start": {
                        "description": "Année de début",
                        "type": "integer"
                    },
                    "year_end": {
                        "description": "Année de fin",
                        "type": "integer"
                    },
                    "uuid": {
                        "description": "UUID unique",
                        "type": "string",
                        "format": "uuid"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ReclamationPeriod": {
                "title": "Reclamation Period",
                "description": "Modèle représentant une période de réclamation",
                "properties": {
                    "id": {
                        "description": "ID unique",
                        "type": "integer"
                    },
                    "academic_year_id": {
                        "description": "ID de l'année académique",
                        "type": "integer"
                    },
                    "start_date": {
                        "description": "Date de début",
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_date": {
                        "description": "Date de fin",
                        "type": "string",
                        "format": "date-time"
                    },
                    "is_active": {
                        "description": "Si la période est active",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    },
                    "academic_year": {
                        "$ref": "#/components/schemas/AcademicYear"
                    }
                },
                "type": "object"
            },
            "SubmissionPeriod": {
                "title": "Submission Period",
                "description": "Modèle représentant une période de soumission",
                "properties": {
                    "id": {
                        "description": "ID unique",
                        "type": "integer"
                    },
                    "academic_year_id": {
                        "description": "ID de l'année académique",
                        "type": "integer"
                    },
                    "start_date": {
                        "description": "Date de début",
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_date": {
                        "description": "Date de fin",
                        "type": "string",
                        "format": "date-time"
                    },
                    "is_active": {
                        "description": "Si la période est active",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    },
                    "academic_year": {
                        "$ref": "#/components/schemas/AcademicYear"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}